🎨 Create svg folder to separate them from pother assets and adjust paths
