+++ src/app/components/Buttons/IconButton.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styles from './IconButton.module.css';
-import SwapIcon from '../assets/SwapIcon';
-import CallIcon from '../assets/CallIcon';
-import MailIcon from '../assets/MailIcon';
-import TrashbinIcon from '../assets/TrashbinIcon';
+import SwapIcon from '../svg/SwapIcon';
+import CallIcon from '../svg/CallIcon';
+import MailIcon from '../svg/MailIcon';
+import TrashbinIcon from '../svg/TrashbinIcon';
 
 export type IconButtonProps = {
   iconType: 'swap' | 'call' | 'mail' | 'trashbin';
