+++ src/app/components/Counter/Counter.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './Counter.module.css';
-import CheckedIcon from '../assets/CheckedIcon';
+import CheckedIcon from '../svg/CheckedIcon';
 
 export type CounterProps = {
   counterType: '1' | '2' | '3' | '4' | '5' | '5+' | '?' | 'checked';
