+++ src/app/components/svg/AddIcon.tsx
