+++ src/app/components/svg/CheckedIcon.tsx
