+++ src/app/components/svg/ConditionIcon.tsx
