+++ src/app/components/svg/HandshakeIcon.tsx
