+++ src/app/components/svg/MailIcon.tsx
