+++ src/app/components/svg/PreviousIcon.tsx
