+++ src/app/components/svg/SwapIcon.tsx
