+++ src/app/components/svg/TrashbinIcon.tsx
