+++ src/app/components/svg/ValueIcon.tsx
